Fixed issue with nuget stdout handling
Using the Start-Process cmdlet with -RedirectStandardOutput does not
reliably flush output to the file immediately.  This causes the text
file that the script immediately tries to read afterwards to sometimes
be empty.  This manifests as the "Execution of NuGet not detected" error
and I suspect is the cause of this issue on certain systems that many
users are seeing.