Handle failures when running NuGet (exitCode != 0, output to stderr)
Previously the script Run-NuGet.ps1 ignored the process exit code so that errors were not propagated to the caller.  This change throws an error when either the process exits with a non-zero value or if errors are written to stderr, which makes Chocolatey more suitable for being called from other scripts.